@ccoffrin 2023-01-05
Nice!

These `basic` functions were designed to be more of a tutorial in nature for teaching the basics of power systems. Would it still get the performance boost if you move this implementation approach into the general `calc_susceptance_matrix_inv` function?  If so, this may be a more suitable location.

https://github.com/lanl-ansi/PowerModels.jl/blob/master/src/core/admittance_matrix.jl#L139-L166

Side note, for very large problems (say 10k or more buses) I find it is even faster to compute the PTDF cuts row-by-row, so you never have to compute the full inverse,

https://github.com/lanl-ansi/PowerModels.jl/blob/master/src/util/flow_limit_cuts.jl#L87-L194