@mtanneau 2023-01-05
>  Would it still get the performance boost if you move this implementation approach into the general calc_susceptance_matrix_inv function?

Let me check

> I find it is even faster to compute the PTDF cuts row-by-row

I have another piece of code to compute PTDF rows lazily, which itself is ~100x faster than what I wrote here (but it needs a cached factorization, so it's less user-friendly).