@mtanneau 2023-01-05
> Would it still get the performance boost if you move this implementation approach into the general calc_susceptance_matrix_inv function?

Partly. Just replacing the two-level loop
https://github.com/lanl-ansi/PowerModels.jl/blob/cf37affe376358592d83851a60bc0270b79e0a4e/src/core/data_basic.jl#L343-L352
with a matrix-matrix product
```julia
B = calc_basic_branch_susceptance_matrix(data)
ptdf = B * b_inv
```
gives a ~5x speedup, which brings it within a factor 2-3 of my current code.

If I also replace the `calc_susceptance_matrix_inv` function to use a more efficient factorization (instead of dense inverse computation), then the computing times become similar.