@ccoffrin 2023-01-06
I like your proposal for updating `calc_susceptance_matrix_inv` and then doing,
```
B = calc_basic_branch_susceptance_matrix(data)
ptdf = B * b_inv
```
I think it will keep this example simple and give the improved performance you are looking for in a broader collection of use cases.