@ccoffrin 2023-01-06
Tests are passing so I am going to merge this.  House keeping before merge,
- Add a note the changelog
- Add your name to the README file as a contributor
- (optional) Add 1-2 sentences of documentation to the `calc_susceptance_matrix_inv` function on what this computational approach is and why in this context it is better than the Julia default / naive approach I had previously used.